requirements.txt
asgiref==3.2.5
autopep8==1.5
Django==3.0.4
django-cors-headers==3.2.1
djangorestframework==3.11.0
gunicorn==20.0.4
Pillow==7.1.1
psycopg2-binary==2.8.4
pycodestyle==2.5.0
python-dotenv==0.12.0
pytz==2019.3
sqlparse==0.3.1
